Add files via upload
Library management system is a typical management Information system (MIS), its Development include the establishment and maintenance of back-end access databases and front-end application development aspects. For the former require the establishment of data consistency and integrity of the strong data security and good libraries. As for the latter, it requires the application fully functional, easy to use and so on. 
Library management system is a project which aims in developing a computerized system to maintain all the daily work of library. This project has many features which are generally not available in normal library management systems like facility of user login. It also has a facility of admin login through which the admin can monitor the whole system.  The librarian after logging into his account i.e. operator’s account can generate various reports such as transaction report, issue report, stock and book report.   
Overall this project of ours is being developed to help the students as well as staff of library to maintain the library in the best way possible and also reduce the human efforts.